Fix style, move things around so test utils are in test dir.
