+++ src/smith/differentiable_numerics/tests/CMakeLists.txt
@@ -4,14 +4,24 @@
 #
 # SPDX-License-Identifier: (BSD-3-Clause)
 
-set(differentiable_numerics_test_depends smith_physics smith_differentiable_numerics gtest)
+set(differentiable_numerics_test_headers
+    differentiable_test_utils.hpp
+    )
 
-set(differentiable_numerics_tests
+smith_add_library(
+    NAME        differentiable_numerics_test_utils
+    HEADERS     ${differentiable_numerics_test_headers}
+    )
+
+set(differentiable_numerics_test_depends smith_physics smith_differentiable_numerics differentiable_numerics_test_utils gtest)
+
+set(differentiable_numerics_test_source
     test_field_state.cpp
     test_explicit_dynamics.cpp
     test_solid_mechanics_state_advancer.cpp
+    differentiable_test_utils.cpp
     )
 
-smith_add_tests( SOURCES       ${differentiable_numerics_tests}
+smith_add_tests( SOURCES       ${differentiable_numerics_test_source}
                  DEPENDS_ON    ${differentiable_numerics_test_depends}
                  NUM_MPI_TASKS 1)
