+++ src/smith/differentiable_numerics/tests/differentiable_test_utils.cpp
@@ -5,7 +5,7 @@
 // SPDX-License-Identifier: (BSD-3-Clause)
 
 #include "gretl/data_store.hpp"
-#include "smith/differentiable_numerics/differentiable_test_utils.hpp"
+#include "smith/differentiable_numerics/tests/differentiable_test_utils.hpp"
 
 namespace smith {
