+++ src/smith/differentiable_numerics/time_discretized_weak_form.hpp
@@ -70,7 +70,7 @@ class SecondOrderTimeDiscretizedWeakForms {
   std::shared_ptr<WeakForm> time_discretized_weak_form;  ///< this publically available abstract weak form is a
                                                          ///< functions of the current u, u_old, v_old, and a_old,
   std::shared_ptr<WeakForm> quasi_static_weak_form;      ///< this publically available abstract weak form is structly a
-                                                         ///< function of the current u, v, and a (no time discretization)
+                                                     ///< function of the current u, v, and a (no time discretization)
 };
 
 template <int spatial_dim, typename OutputSpace, typename inputs = Parameters<>>
